program now reads file and prints anagrams
I've managed to make the program ignore capital letters while looking
for and printing anagrams and the original word, but I've only managed
to get half way there.

For example in the last run I recieved: "Edna edna".  they are the same
word and therfore should not print.  But on a previous run I got: "Edna
dean".

As I said, half way there...